pyproject.toml
[project]
name = "flowbite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastwand>=0.1.31",
    "monsterui>=1.0.10",
    "python-fasthtml>=0.12.1",
]
